+++ backend/src/main/java/com/tripfriend/domain/qna/service/QuestionService.java
@@ -4,8 +4,13 @@ import com.tripfriend.domain.member.member.entity.Member;
 import com.tripfriend.domain.member.member.repository.MemberRepository;
 import com.tripfriend.domain.qna.entity.Question;
 import com.tripfriend.domain.qna.repository.QuestionRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 import java.util.List;
 
@@ -16,10 +21,7 @@ public class QuestionService {
     private final MemberRepository memberRepository;
 
     //질문 생성
-    public Question createQuestion(String title, String content, Long memberId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
-
+    public Question createQuestion(String title, String content, Member member) {
         Question question = Question.builder()
                 .title(title)
                 .content(content)
@@ -35,14 +37,27 @@ public class QuestionService {
 
     //특정 질문 조회
     public Question getQuestionById(Long id) {
-        return questionRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("해당 질문을 찾을 수 없습니다."));
+        return questionRepository.findByIdWithMember(id)
+                .orElseThrow(() -> new RuntimeException("해당 질문을 찾을 수 없습니다."));
     }
 
     //질문 삭제
-    public void deleteQuestionById(Long id) {
+    public void deleteQuestionById(Long id, Member member) {
         Question question = getQuestionById(id);
+
+        // 작성자인지 확인
+        if (!question.getMember().getId().equals(member.getId())) {
+            throw new IllegalArgumentException("작성자만 질문을 삭제할 수 있습니다.");
+        }
+
         questionRepository.delete(question);
     }
 
-}
+
+
+
+    }
+
+
+
+
